apps/api/pyproject.toml
[project]
name = "regknots-api"
version = "0.1.0"
description = "RegKnots API server"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "asyncpg>=0.30.0",
    "redis>=5.0.0",
    "pydantic-settings>=2.6.0",
    # Migrations
    "alembic>=1.14.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    # Auth
    "pyjwt>=2.10.0",
    "pwdlib[argon2]>=0.2.0",
    "email-validator>=2.2.0",
    # RAG
    "anthropic>=0.49.0",
    "openai>=1.0.0",
    "tiktoken>=0.7.0",
    "regknots-rag",
]

[tool.uv.sources]
regknots-rag = { path = "../../packages/rag", editable = true }

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]
